+++ src/models/constants/defaultCookieName.ts
@@ -1,0 +1,1 @@
+export const defaultCookieConsentName = "CookieConsent";