+++ src/models/constants/sameSiteOptions.ts
@@ -1,0 +1,5 @@
+export enum SAME_SITE_OPTIONS {
+  STRICT = "strict",
+  LAX = "lax",
+  NONE = "none",
+}