+++ stories/defaults/intro.tsx
@@ -1,0 +1,11 @@
+import { FunctionComponent } from "react";
+
+type Props = {};
+
+export const Intro: FunctionComponent<Props> = () => {
+  return (
+    <h1>
+      <p>Debug is turned on for all stories so that the bar always shows up</p>
+    </h1>
+  );
+};