+++ stories/defaults/storyProps.ts
@@ -1,0 +1,3 @@
+import { CookieConsentProps, defaultCookieConsentProps } from "../../src/CookieConsent.props";
+
+export const defaultStoryProps: CookieConsentProps = { ...defaultCookieConsentProps, debug: true };