+++ stories/stories/overlay.story.tsx
@@ -1,0 +1,11 @@
+import { CookieConsentProps } from "../../src/CookieConsent.props";
+import { defaultStoryProps } from "../defaults/storyProps";
+import { DefaultTemplate } from "../defaults/template";
+
+const Overlay = DefaultTemplate.bind({});
+Overlay.args = {
+  ...defaultStoryProps,
+  overlay: true,
+} as CookieConsentProps;
+
+export { Overlay };